update monoduck to v0.2.0; useStore and jsend

+++ src/backend/controllers/flag-inapi.ts
@@ -1,4 +1,4 @@
-import { _, tapiduck, TapiError } from 'monoduck'
+import { _, tapiduck } from 'monoduck'
 import type { ZFlag } from '../../shared/z-models'
 import { zModeEnum, defaultFlagRow } from '../../shared/z-models'
 import { api } from '../../shared/endpoints'
@@ -8,12 +8,15 @@ import { auth } from '../auth'
 import { getModeRing } from '../../shared/helpers'
 import { emitFlagNotif } from './sock-util'
 
-tapiduck.route(app, api.internal.createFlag, async function (reqdata) {
+tapiduck.route(app, api.internal.createFlag, async function (reqdata, jsend) {
   const me = await auth.getMe(reqdata.inapiToken)
+  if (!me) {
+    return jsend.fail(auth.generalFailText)
+  }
   const flagId = reqdata.flag_id
   const existingFlag = await models.flag.findById(flagId)
   if (_.bool(existingFlag)) {
-    throw new TapiError(`Flag with ID '${flagId}' already exists.`)
+    return jsend.fail(`Flag with ID '${flagId}' already exists.`)
   }
   const flag: ZFlag = {
     ...defaultFlagRow,
@@ -25,19 +28,26 @@ tapiduck.route(app, api.internal.createFlag, async function (reqdata) {
   _.each(zModeEnum.options, function (mode) {
     emitFlagNotif(flag.id, mode)
   })
-  return flag
+  return jsend.success(flag)
 })
 
-tapiduck.route(app, api.internal.getFlags, async function (reqdata) {
-  await auth.getMe(reqdata.inapiToken)
-  return await models.flag.findAll({})
-})
-
-tapiduck.route(app, api.internal.updateFlag, async function (reqdata) {
+tapiduck.route(app, api.internal.getFlags, async function (reqdata, jsend) {
   const me = await auth.getMe(reqdata.inapiToken)
+  if (!me) {
+    return jsend.fail(auth.generalFailText)
+  }
+  const flags: ZFlag[] = await models.flag.findAll({})
+  return jsend.success(flags)
+})
+
+tapiduck.route(app, api.internal.updateFlag, async function (reqdata, jsend) {
+  const me = await auth.getMe(reqdata.inapiToken)
+  if (!me) {
+    return jsend.fail(auth.generalFailText)
+  }
   const oldFlag = await models.flag.findOne({ where: { id: reqdata.flag.id } })
   if (_.not(oldFlag)) {
-    throw new TapiError('No such flag.')
+    return jsend.fail('No such flag.')
   }
   const updatedFlag: ZFlag = {
     ...oldFlag,
@@ -52,18 +62,21 @@ tapiduck.route(app, api.internal.updateFlag, async function (reqdata) {
       emitFlagNotif(updatedFlag.id, mode)
     }
   })
-  return updatedFlag
+  return jsend.success(updatedFlag)
 })
 
-tapiduck.route(app, api.internal.deleteFlag, async function (reqdata) {
-  await auth.getMe(reqdata.inapiToken)
+tapiduck.route(app, api.internal.deleteFlag, async function (reqdata, jsend) {
+  const me = await auth.getMe(reqdata.inapiToken)
+  if (!me) {
+    return jsend.fail(auth.generalFailText)
+  }
   const flag = await models.flag.findOne({ where: { id: reqdata.flag_id } })
   if (_.not(flag)) {
-    throw new TapiError('No such flag.')
+    return jsend.fail('No such flag.')
   }
   // Delete the flag, along with all linked rules (in either mode).
   const rules = await models.rule.findAll({ where: { flag_id: flag.id } })
-  // TODO:? _Consider_ deleting multiple flags in a single DB roundtrip.
+  // TODO:? _Consider_ deleting multiple rules in a single DB roundtrip.
   await Promise.all(rules.map(async r => await models.rule.deleteById(r.id)))
   await models.flag.deleteById(flag.id)
   _.each(zModeEnum.options, function (mode) {
@@ -73,5 +86,5 @@ tapiduck.route(app, api.internal.deleteFlag, async function (reqdata) {
       emitFlagNotif(flag.id, mode)
     }
   })
-  return { id: flag.id }
+  return jsend.success({ id: flag.id })
 })
